Fixed bug where modal wouldn't show list of maps on initial render; for slow connections it will put up a loading message for the user, also removed utils.js for now as the getMaps fn in it wasn't needed anymore as the code was moved into select_map_modal
