+++ blurp-app/src/components/select_map_modal.jsx
@@ -7,7 +7,6 @@ import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
 import IconButton from '@mui/material/IconButton';
 import { BACKEND_URL, MODAL_VIEW } from '../constants/constants';
-import getMaps from '../utils/utils';
 
 /*
 Pulled the modal from an example here:
@@ -16,12 +15,61 @@ https://codesandbox.io/s/nw2r1e?file=/demo.tsx:221-423
 */
 
 const LoadMapModal = forwardRef((props, ref) => {
-  const [maps, setMaps] = useState([]);
+  const [maps, setMaps] = useState();
+  const [isLoading, setLoading] = useState(true);
   const [mapName, setMapName] = useState('');
 
   useEffect(() => {
-    setMaps(getMaps(props.profile));
-  }, [props.modal]);
+    const id = props.profile.userID;
+    let list = [];
+    axios
+      .post(BACKEND_URL + '/map', {
+        userID: id,
+      })
+      .then((response) => {
+        response.data.forEach((current) => {
+          list.push([current.mapID, current.title]);
+        });
+        setMaps(
+          list.map((value) => (
+            <ListItem
+              key={value[0]}
+              disableGutters
+              sx={{ width: '98%' }}
+              className="my-3 rounded-lg bg-gray-50"
+              secondaryAction={
+                <div className="h-full">
+                  <button
+                    value={value[0]}
+                    className="h-full rounded-l-lg bg-green-600 p-2 font-bold text-white hover:bg-green-900"
+                    onClick={(e) => {
+                      props.changeProfile(props.profile.userID, e.target.value, true);
+
+                      ref.current.LoadFromDB(e.target.value);
+                      handleClose();
+                    }}>
+                    Select
+                  </button>
+                  <button
+                    value={value[0]}
+                    className="h-full rounded-r-lg bg-red-600 p-2 font-bold text-white hover:bg-red-900"
+                    onClick={(e) => {
+                      deleteMap(props.profile, e.target.value);
+                    }}>
+                    Delete
+                  </button>
+                </div>
+              }>
+              <div className="mx-2">{value[1]}</div>
+            </ListItem>
+          ))
+        );
+        setLoading(false);
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  }, []);
 
   const handleClose = (reason) => {
     /* // Prevents the user from clicking the outside of the modal on start
@@ -70,6 +118,16 @@ const LoadMapModal = forwardRef((props, ref) => {
       });
   };
 
+  const mapList = () => {
+    if (!isLoading) {
+      return <>{maps}</>;
+    } else {
+      <>
+        <h1>Loading...</h1>
+      </>;
+    }
+  };
+
   const selectView = (view) => {
     switch (view) {
       case MODAL_VIEW.START:
@@ -79,38 +137,7 @@ const LoadMapModal = forwardRef((props, ref) => {
             <List
               sx={{ margin: '12px 0 12px 0' }}
               className="grid max-h-96 w-full grid-cols-1 justify-items-center overflow-auto rounded-lg bg-neutral-400">
-              {maps.map((value) => (
-                <ListItem
-                  key={value[0]}
-                  disableGutters
-                  sx={{ width: '98%' }}
-                  className="my-3 rounded-lg bg-gray-50"
-                  secondaryAction={
-                    <div className="h-full">
-                      <button
-                        value={value[0]}
-                        className="h-full rounded-l-lg bg-green-600 p-2 font-bold text-white hover:bg-green-900"
-                        onClick={(e) => {
-                          props.changeProfile(props.profile.userID, e.target.value, true);
-
-                          ref.current.LoadFromDB(e.target.value);
-                          handleClose();
-                        }}>
-                        Select
-                      </button>
-                      <button
-                        value={value[0]}
-                        className="h-full rounded-r-lg bg-red-600 p-2 font-bold text-white hover:bg-red-900"
-                        onClick={(e) => {
-                          deleteMap(props.profile, e.target.value);
-                        }}>
-                        Delete
-                      </button>
-                    </div>
-                  }>
-                  <div className="mx-2">{value[1]}</div>
-                </ListItem>
-              ))}
+              {mapList()}
             </List>
             <form>
               <div>
